backend/requirements.txt
fastapi~=0.75.2
pandas~=1.3.4
requests~=2.27.1
python-multipart==0.0.5
uvicorn~=0.17.6
lightgbm==3.2.1
pyyaml~=6.0
scikit-learn~=1.0.2
optuna~=3.1.0
joblib~=1.1.0
numpy~=1.20.3